Move all code in the single StoredKey.swift file into multiple files in a StoredKey/ sub-directory
