+++ forApple/CaptiveCrypto/CaptiveCrypto/Array+inserting.swift
@@ -1,0 +1,12 @@
+// Copyright 2023 VMware, Inc.
+// SPDX-License-Identifier: BSD-2-Clause
+
+import Foundation
+
+extension Array {
+    func inserting(_ element:Element, at index:Int) -> Array<Element> {
+        var inserted = self
+        inserted.insert(element, at: index)
+        return inserted
+    }
+}
