+++ forApple/CaptiveCrypto/CaptiveCrypto/StoredKey/Storage.swift
@@ -1,0 +1,25 @@
+// Copyright 2023 VMware, Inc.
+// SPDX-License-Identifier: BSD-2-Clause
+
+import Foundation
+
+extension StoredKey {
+    // Enumeration for different storage of keys supported by this class, either
+    // as generic passwords in the keychain, or as keys in the keychain. The
+    // keychain only stores private keys as keys, so symmetric keys must be
+    // stored as generic passwords.
+    enum Storage: String, CaseIterable {
+        case generic, key
+
+        var secClass:CFString {
+            switch self {
+            case .generic:
+                return kSecClassGenericPassword
+            case .key:
+                return kSecClassKey
+            }
+        }
+    }
+    
+
+}
