+++ forApple/CaptiveCrypto/CaptiveCrypto/StoredKey/StoredKey.swift
@@ -1,0 +1,44 @@
+// Copyright 2023 VMware, Inc.
+// SPDX-License-Identifier: BSD-2-Clause
+
+import Foundation
+import CryptoKit
+
+class StoredKey {
+    // Properties and methods of a StoredKey instance. It isn't necessary to use
+    // StoredKey instances externally. The static methods, like
+    // encypt(message withFirstKeyNamed:) for example, can be used instead.
+    internal let _storage:Storage
+    let secKey:SecKey?
+    let symmetricKey:SymmetricKey?
+
+    var storage:String {_storage.rawValue}
+
+    // Symmetric key constructor.
+    init(_ symmetricKey:SymmetricKey) {
+        _storage = .generic
+        secKey = nil
+        self.symmetricKey = symmetricKey
+    }
+    
+    // Key pair constructor. The SecKey will be the private key. The
+    // corresponding private key will be generated if needed in the
+    // encryptBasedOnPrivateKey() method, below.
+    init(_ secKey:SecKey) {
+        _storage = .key
+        self.secKey = secKey
+        symmetricKey = nil
+    }
+
+}
+
+protocol StoredKeyBasis {
+    func storedKey() -> StoredKey
+}
+extension SecKey: StoredKeyBasis {
+    func storedKey() -> StoredKey { return StoredKey(self) }
+}
+extension SymmetricKey: StoredKeyBasis {
+    func storedKey() -> StoredKey { return StoredKey(self) }
+}
+
